Criação dos dois usuários: Seguidor e Seguido
