--- semana-18/ts-jest-boilerplate-final/ts-jest-boilerplate-final/src/business/usecases/user/FollowUser.ts
@@ -1,23 +1,0 @@
-import { UserGateway } from '../../gateways/user/UserGateway';
-
-export class FollowUserUseCase {
-    constructor( private userGateway: UserGateway){}
-
-    async execute(input: FollowUserInput) {
-        await this.verifyUserExists(input)
-
-        //await this.userGateway.createUserRelation(input.followerId, input.followedId)
-    }
-
-    async verifyUserExists(input: FollowUserInput) {
-        await Promise.all([
-            this.userGateway.verifyUserExists(input.followedId),
-            this.userGateway.verifyUserExists(input.followerId)
-        ])
-    }
-}
-
-export interface FollowUserInput {
-    followerId: string,
-    followedId: string
-}
